Move url-controls sync logic to the _page.py and page.js
